+++ praetorian_cli/handlers/test.py
@@ -20,6 +20,4 @@ def test(chariot, key, suite):
         command.extend(['-k', key])
     if suite:
         command.extend(['-m', suite])
-    # REMOVE
-    command.extend(['-s'])
     pytest.main(command)